+++ spec/components/phase_banner_spec.rb
@@ -1,0 +1,43 @@
+require 'rails_helper'
+
+describe "Phase banner", type: :view do
+  def component_name
+    "phase_banner"
+  end
+
+  it "errors without a phase" do
+    expect {
+      render_component({})
+    }.to raise_error("The phase banner component expects a `phase` (`beta` or `alpha`), nil given")
+  end
+
+  it "does not error without a message" do
+    render_component(phase: "alpha")
+
+    assert_select ".gem-c-phase-banner"
+  end
+
+  it "shows a custom message" do
+    render_component(phase: "alpha", message: "custom message")
+
+    assert_select ".gem-c-phase-banner span", text: "custom message"
+  end
+
+  it "allows custom message HTML" do
+    render_component(phase: "beta", message: "custom <strong>message</strong>".html_safe)
+
+    assert_select ".gem-c-phase-banner strong", text: "message"
+  end
+
+  it "shows the alpha phase" do
+    render_component(phase: "alpha")
+
+    assert_select ".phase-tag", text: "alpha"
+  end
+
+  it "shows the beta phase" do
+    render_component(phase: "beta")
+
+    assert_select ".phase-tag", text: "beta"
+  end
+end